--- pyonsnalcolor-api/src/main/java/com/pyonsnalcolor/push/fcm/FcmController.java
@@ -1,24 +1,0 @@
-package com.pyonsnalcolor.push.fcm;
-
-import com.pyonsnalcolor.push.dto.DeviceTokenRequestDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-import java.util.concurrent.ExecutionException;
-
-@Controller
-@RequiredArgsConstructor
-public class FcmController {
-
-    private final FcmPushService fcmPushService;
-
-    // ios 테스트용
-    @GetMapping("/fcm/test")
-    public ResponseEntity fcmTest(DeviceTokenRequestDto deviceTokenRequestDto) throws ExecutionException, InterruptedException {
-        fcmPushService.pbFcmTest(deviceTokenRequestDto);
-        return new ResponseEntity(HttpStatus.OK);
-    }
-}
